+++ src/main/java/roomescape/reservation/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
